test: move `wait_task_completion` function in `utils.py` for better reuse
