+++ tests/integration/test_studies_upgrade.py
@@ -1,34 +1,11 @@
 import os
-import time
 
 import pytest
+from antarest.core.tasks.model import TaskStatus
 from starlette.testclient import TestClient
-
-from antarest.core.tasks.model import TaskDTO, TaskStatus
+from tests.integration.utils import wait_task_completion
 
 RUN_ON_WINDOWS = os.name == "nt"
-
-
-def wait_task_completion(
-    client: TestClient,
-    access_token: str,
-    task_id: str,
-    *,
-    timeout: float = 10,
-) -> TaskDTO:
-    end_time = time.time() + timeout
-    while time.time() < end_time:
-        time.sleep(0.1)
-        res = client.get(
-            f"/v1/tasks/{task_id}",
-            headers={"Authorization": f"Bearer {access_token}"},
-            json={"wait_for_completion": True},
-        )
-        assert res.status_code == 200
-        task = TaskDTO(**res.json())
-        if task.status not in {TaskStatus.PENDING, TaskStatus.RUNNING}:
-            return task
-    raise TimeoutError(f"{timeout} seconds")
 
 
 class TestStudyUpgrade:
